+++ Src/TinyXML/tinystr.cpp
@@ -36,7 +36,7 @@ distribution.
 // TiXmlString constructor, based on a C string
 TiXmlString::TiXmlString (const char* instring)
 {
-    unsigned newlen;
+    size_t newlen;
     char * newstring;
 
     if (!instring)
@@ -58,7 +58,7 @@ TiXmlString::TiXmlString (const char* instring)
 // TiXmlString copy constructor
 TiXmlString::TiXmlString (const TiXmlString& copy)
 {
-    unsigned newlen;
+    size_t newlen;
     char * newstring;
 
 	// Prevent copy to self!
@@ -84,7 +84,7 @@ TiXmlString::TiXmlString (const TiXmlString& copy)
 // TiXmlString = operator. Safe when assign own content
 void TiXmlString ::operator = (const char * content)
 {
-    unsigned newlen;
+    size_t newlen;
     char * newstring;
 
     if (! content)
@@ -105,15 +105,15 @@ void TiXmlString ::operator = (const char * content)
 // = operator. Safe when assign own content
 void TiXmlString ::operator = (const TiXmlString & copy)
 {
-    unsigned newlen;
+    size_t newlen;
     char * newstring;
 
-    if (! copy . length ())
+    if (!copy.length())
     {
-        empty_it ();
+        empty_it();
         return;
     }
-    newlen = copy . length () + 1;
+    newlen = copy.length() + 1;
     newstring = new char [newlen];
     // strcpy (newstring, copy . c_str ());
     memcpy (newstring, copy . c_str (), newlen);
@@ -125,10 +125,10 @@ void TiXmlString ::operator = (const TiXmlString & copy)
 
 
 // append a const char * to an existing TiXmlString
-void TiXmlString::append( const char* str, int len )
+void TiXmlString::append( const char* str, ssize_t len )
 {
     char * new_string;
-    unsigned new_alloc, new_size, size_suffix;
+    size_t new_alloc, new_size, size_suffix;
 	
 	// don't use strlen - it can overrun the len passed in!
 	const char* p = str;
@@ -189,7 +189,7 @@ void TiXmlString::append( const char* str, int len )
 void TiXmlString::append( const char * suffix )
 {
     char * new_string;
-    unsigned new_alloc, new_size;
+    size_t new_alloc, new_size;
 
     new_size = length () + strlen (suffix) + 1;
     // check if we need to expand
@@ -247,7 +247,7 @@ void TiXmlString::append( const char * suffix )
 //}
 
 
-unsigned TiXmlString::find (char tofind, unsigned offset) const
+size_t TiXmlString::find (char tofind, size_t offset) const
 {
     char * lookup;
